:bug: debug in display each threshold and distribution

+++ dataset/hpa_dataset.py
@@ -213,7 +213,8 @@ class HPAData(data.Dataset):
         self.id_to_indices = {v: k for k, v in self.indices_to_id.items()}
 
         if self.writer:
-            data_dict = np.array(self.dataframe['Target']).astype(np.byte)
+            data_dict = np.array(self.dataframe['Target'])
+            data_dict = data_dict.astype(np.byte)
             data_dict = np.bincount(data_dict.flatten())
             F = plt.figure()
             plt.bar(list(range(len(data_dict))), data_dict)
